@@ -2,7 +2,7 @@
   <Import Project="..\nuget.props" />
 
   <PropertyGroup>
-    <TargetFrameworks>netcoreapp3.0;netcoreapp2.1;netstandard2.1;netstandard2.0;net471;net451</TargetFrameworks>
+    <TargetFrameworks>netcoreapp3.1;netcoreapp2.1;netstandard2.1;netstandard2.0;net471;net451</TargetFrameworks>
     <RootNamespace>SpanJson</RootNamespace>
     <AssemblyName>SpanJson.Extensions</AssemblyName>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
@@ -20,9 +20,6 @@
   <ItemGroup>
     <PackageReference Include="CuteAnt.Runtime" Version="$(CuteAntRuntimePackageVersion)" />
     <PackageReference Include="Newtonsoft.Json" Version="$(NewtonsoftJsonPackageVersion)" />
-  </ItemGroup>
-  <ItemGroup Condition=" '$(TargetFramework)' != 'netcoreapp3.0' and '$(TargetFramework)' != 'net451' ">
-    <!--<PackageReference Include="System.Text.Json" Version="$(SystemTextJsonPackageVersion)" />-->
   </ItemGroup>
 
   <ItemGroup>
