@@ -1,7 +1,7 @@
 ﻿using System;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ApplicationParts;
-using Microsoft.AspNetCore.Mvc.Internal;
+//using Microsoft.AspNetCore.Mvc.Internal;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.ObjectPool;
@@ -18,8 +18,9 @@ namespace SpanJson.AspNetCore.Formatter.Tests
         {
             var options = GetOptions<MvcOptions>(services =>
             {
-                var builder = new MvcCoreBuilder(services, new ApplicationPartManager());
-                useAction(builder);
+                var mock = new Mock<IMvcCoreBuilder>();
+                mock.Setup(a => a.Services).Returns(services);
+                useAction(mock.Object);
             });
             Assert.NotNull(options);
             Assert.Single(options.InputFormatters);
@@ -32,8 +33,9 @@ namespace SpanJson.AspNetCore.Formatter.Tests
         {
             var options = GetOptions<MvcOptions>(services =>
             {
-                var builder = new MvcBuilder(services, new ApplicationPartManager());
-                useAction(builder);
+                var mock = new Mock<IMvcBuilder>();
+                mock.Setup(a => a.Services).Returns(services);
+                useAction(mock.Object);
             });
             Assert.NotNull(options);
             Assert.Single(options.InputFormatters);
